Fix integer overflow in seekbar calculations
Agent-Logs-Url: https://github.com/9xhk-1/163-music-for-watch/sessions/ac17189f-924b-400c-9638-97bf5321d34d

+++ app/src/main/java/com/watch/music163/MainActivity.java
@@ -85,7 +85,7 @@ public class MainActivity extends AppCompatActivity implements MusicPlayerManage
                 if (fromUser) {
                     int duration = playerManager.getDuration();
                     if (duration > 0) {
-                        tvCurrentTime.setText(formatTime(progress * duration / 1000));
+                        tvCurrentTime.setText(formatTime((int) ((long) progress * duration / 1000)));
                     }
                 }
             }
@@ -99,7 +99,7 @@ public class MainActivity extends AppCompatActivity implements MusicPlayerManage
             public void onStopTrackingTouch(SeekBar bar) {
                 int duration = playerManager.getDuration();
                 if (duration > 0) {
-                    int seekPos = bar.getProgress() * duration / 1000;
+                    int seekPos = (int) ((long) bar.getProgress() * duration / 1000);
                     playerManager.seekTo(seekPos);
                 }
                 isUserSeeking = false;
